+++ src-tauri/src/audio/mod.rs
@@ -10,6 +10,7 @@ mod decode;
 mod dev_io;
 pub mod device_commands;
 pub mod mix_commands;
+mod play_input;
 pub mod preload_commands;
 pub(crate) mod progress_task;
 pub mod radio_commands;
